NAV-22215: Flyttet test fra HjemlerService to BrevUtilKtTest
